feat(client): migrate to TypeScript with Vite
