+++ .eslintignore
@@ -1,3 +1,2 @@
-ttl_client
 ttl_db
 node_modules
