+++ package.json
@@ -9,8 +9,8 @@
     "lint": "ESLINT_USE_FLAT_CONFIG=false eslint .",
     "format": "prettier . --write",
     "build": "tsc -b",
-    "test": "npm run -w ttl_server test",
-    "test:watch": "npm run -w ttl_server test -- --watch",
+    "test": "vitest run",
+    "test:watch": "vitest",
     "coverage": "npm run -w ttl_server coverage",
     "openapi": "npm run -w ttl_server openapi"
   },
@@ -24,6 +24,7 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.30.0",
+    "jsdom": "^26.1.0",
     "prettier": "^3.3.3",
     "supertest": "^7.0.0",
     "tsx": "^4.19.0",
