+++ tsconfig.json
@@ -1,6 +1,7 @@
 {
   "files": [],
   "references": [
-    { "path": "ttl_server" }
+    { "path": "ttl_server" },
+    { "path": "ttl_client" }
   ]
 }
