+++ ttl_client/src/components/OrderLotMain.tsx
@@ -1,33 +1,44 @@
-import React, { useState, useEffect } from 'react';
-import ViewOrder from './ViewOrder';
+import { useEffect, useState } from 'react';
+
 import InputForm from './InputForm';
+import ViewOrder from './ViewOrder';
 
 const regex = /^(\d{7})_(\d{3})_(\d\d-\d\d-\d\d\d\d)_(.*)_(.*)\$$/;
 
-function OrderLotMain({ token }) {
+interface Props {
+  token: string;
+}
+
+interface Lot {
+  lot_id?: string;
+  order_id?: string;
+}
+
+function OrderLotMain({ token }: Props) {
   const [orderID, setOrderID] = useState('');
   const [lotID, setLotID] = useState('');
   const [isCurrentOrderInDb, setIsCurrentOrderInDb] = useState(false);
-  const [currentLot, setCurrentLot] = useState([]);
+  const [currentLot, setCurrentLot] = useState<Lot[]>([]);
   const [isSearchingMode, setIsSearchingMode] = useState(false);
 
-  // Asynchrone functie voor het ophalen van data
-  const fetchOrderData = async (orderID) => {
+  const fetchOrderData = async (id: string) => {
     try {
       const response = await fetch(`api/orders`, {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
-          'Authorization': `Bearer ${token}`,
+          Authorization: `Bearer ${token}`,
         },
-        body: JSON.stringify({ order_id: orderID }),
+        body: JSON.stringify({ order_id: id }),
       });
-      const { is_order_in_db, lot_numbers } = await response.json();
-
+      const { is_order_in_db, lot_numbers } = (await response.json()) as {
+        is_order_in_db: boolean;
+        lot_numbers?: Lot[];
+      };
       setIsCurrentOrderInDb(is_order_in_db);
-      setCurrentLot(lot_numbers || []); // Zorg ervoor dat er geen null-waarden worden verwerkt
+      setCurrentLot(lot_numbers || []);
     } catch (err) {
-      console.error('Fout bij het ophalen van ordergegevens:', err.message);
+      console.error('Fout bij het ophalen van ordergegevens:', (err as Error).message);
       setIsCurrentOrderInDb(false);
       setCurrentLot([]);
     }
@@ -35,12 +46,12 @@ function OrderLotMain({ token }) {
 
   useEffect(() => {
     if (regex.test(orderID)) {
-      fetchOrderData(orderID); // Roep de asynchrone functie aan
+      void fetchOrderData(orderID);
     } else {
       setIsCurrentOrderInDb(false);
       setCurrentLot([]);
     }
-  }, [orderID, token]); // 'token' als dependency toegevoegd
+  }, [orderID, token]);
 
   return (
     <>
@@ -49,21 +60,19 @@ function OrderLotMain({ token }) {
         setOrderID={setOrderID}
         lotID={lotID}
         setLotID={setLotID}
-        currentLot={currentLot}
-        setCurrentLot={setCurrentLot}
+        setCurrentLot={(lot: unknown) => setCurrentLot(lot as Lot[])}
         isCurrentOrderInDb={isCurrentOrderInDb}
         isSearchingMode={isSearchingMode}
         setIsSearchingMode={setIsSearchingMode}
         token={token}
-        />
+      />
       {currentLot.length > 0 && (
         <ViewOrder
-        isCurrentOrderInDb={isCurrentOrderInDb}
-        currentLot={currentLot}
-        orderID={orderID}
-        setCurrentLot={setCurrentLot}
-        setOrderID={setOrderID}
-        token={token}
+          currentLot={currentLot}
+          orderID={orderID}
+          setCurrentLot={setCurrentLot}
+          setOrderID={setOrderID}
+          token={token}
         />
       )}
     </>
