--- ttl_client/src/setupProxy.js
@@ -1,14 +1,0 @@
-const createProxyMiddleware  = require('http-proxy-middleware');
-
-module.exports = function(app) {
-  app.use(
-    '/api/',
-    createProxyMiddleware({
-      target: 'http://server:5000',
-      changeOrigin: true,
-      pathRewrite: {
-        '^/api': '/'
-    }
-    })
-  );
-};
