+++ ttl_client/src/utils/sum.ts
@@ -1,0 +1,3 @@
+export function sum(a: number, b: number): number {
+  return a + b;
+}
