+++ vitest.workspace.ts
@@ -1,0 +1,6 @@
+import { defineWorkspace } from 'vitest/config';
+
+export default defineWorkspace([
+  'ttl_server/vitest.config.ts',
+  'ttl_client/vitest.config.ts'
+]);
